+++ src/mastodon_api/media/views.rs
@@ -1,5 +1,12 @@
 /// https://docs.joinmastodon.org/methods/media/#v1
-use actix_web::{post, web, HttpResponse, Scope};
+use actix_multipart::form::MultipartForm;
+use actix_web::{
+    post,
+    web,
+    Either,
+    HttpResponse,
+    Scope,
+};
 use actix_web_httpauth::extractors::bearer::BearerAuth;
 
 use mitra_config::Config;
@@ -16,15 +23,26 @@ use crate::mastodon_api::{
     uploads::save_b64_file,
 };
 
-use super::types::{AttachmentCreateData, Attachment};
+use super::types::{
+    Attachment,
+    AttachmentData,
+    AttachmentDataMultipartForm
+};
 
 #[post("")]
 async fn create_attachment_view(
     auth: BearerAuth,
     config: web::Data<Config>,
     db_pool: web::Data<DbPool>,
-    attachment_data: web::Json<AttachmentCreateData>,
+    attachment_data: Either<
+        MultipartForm<AttachmentDataMultipartForm>,
+        web::Json<AttachmentData>,
+    >,
 ) -> Result<HttpResponse, MastodonError> {
+    let attachment_data = match attachment_data {
+        Either::Left(form) => form.into_inner().into(),
+        Either::Right(data) => data.into_inner(),
+    };
     let db_client = &**get_database_client(&db_pool).await?;
     let current_user = get_current_user(db_client, auth.token()).await?;
     let media_storage = MediaStorage::from(config.as_ref());